+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { Component, StrictMode } from 'react';
 import PropTypes from 'prop-types';
-import { okapi as okapiConfig, config } from 'stripes-config';
+import { okapi as localOkapi, branding as localBranding, config as localConfig } from 'stripes-config';
+import isEmpty from 'lodash/isEmpty';
 import merge from 'lodash/merge';
 import localforage from 'localforage';
 import AppConfigError from './components/AppConfigError';
@@ -19,8 +20,8 @@ import { eventsPortal, stripesHubAPI } from './constants';
 import { getLoginTenant } from './loginServices';
 
 
-const StrictWrapper = ({ children }) => {
-  if (config.disableStrictMode) {
+const StrictWrapper = ({ children, config }) => {
+  if (config?.disableStrictMode) {
     return children;
   }
 
@@ -60,8 +61,59 @@ export const isStorageEnabled = () => {
   return isEnabled;
 };
 
+export const getStripesHubConfig = () => {
+  try {
+    const folioConfig = JSON.parse(localStorage.getItem(stripesHubAPI.FOLIO_CONFIG_KEY) || '{}');
+    const brandingConfig = JSON.parse(localStorage.getItem(stripesHubAPI.BRANDING_CONFIG_KEY) || '{}');
+    return { folioConfig, brandingConfig };
+  } catch (error) {
+    console.error('Failed to parse StripesHub config from localStorage:', error); // eslint-disable-line no-console
+    // If there was an error parsing the config from StripesHub, return empty objects so that we fall back to stripes.config.js values.
+    return { folioConfig: {}, brandingConfig: {} };
+  }
+};
+
+/**
+ * If StripesHub is present in localstorage, override config values as StripesHub is now the source of truth.
+ *
+ * @param {object} theLocalOkapi the Okapi config object from stripes.config.js.
+ * @param {object} theLocalConfig the Stripes config object from stripes.config.js.
+ * @param {object} theLocalBranding the branding config from stripes.config.js.
+ * @param {object} stripesHub the config object from StripesHub.
+ * @returns Updated config object with values from Stripes Hub if available, or default to stripes.config.js
+ */
+export const getOverrideConfig = (theLocalOkapi, theLocalConfig, theLocalBranding, stripesHub) => {
+  let stripesConfig = {};
+  let stripesOkapi = {};
+  let stripesBranding = {};
+
+  // If folioConfig is present in StripesHub, use it to override values from stripes.config.js.
+  // Otherwise, use values from stripes.config.js.
+  if (stripesHub && !isEmpty(stripesHub.folioConfig)) {
+    // Pass URLs from FOLIO config into Okapi config.
+    stripesOkapi.url = stripesHub.folioConfig.gatewayUrl;
+    stripesOkapi.authnUrl = stripesHub.folioConfig.authnUrl;
+
+    // Pass all FOLIO config values from StripesHub config into Stripes config.
+    stripesConfig = stripesHub.folioConfig;
+
+    // Remove URLs that StripesHub consolidates into FOLIO config, but classic Stripes expects to find in Okapi config.
+    delete stripesConfig.gatewayUrl;
+    delete stripesConfig.authnUrl;
+
+    stripesBranding = stripesHub.brandingConfig;
+  } else {
+    stripesOkapi = theLocalOkapi;
+    stripesConfig = theLocalConfig;
+    stripesBranding = theLocalBranding;
+  }
+
+  return { stripesOkapi, stripesConfig, stripesBranding };
+};
+
 StrictWrapper.propTypes = {
   children: PropTypes.node.isRequired,
+  config: PropTypes.object.isRequired,
 };
 
 export default class StripesCore extends Component {
@@ -74,14 +126,18 @@ export default class StripesCore extends Component {
     super(props);
 
     if (isStorageEnabled()) {
-      const parsedTenant = getLoginTenant(okapiConfig, config);
+      const stripesHubConfig = getStripesHubConfig();
+      const { stripesOkapi, stripesConfig, stripesBranding } = getOverrideConfig(localOkapi, localConfig, localBranding, stripesHubConfig);
+      const parsedTenant = getLoginTenant(localOkapi, stripesConfig);
 
-      const okapi = (typeof okapiConfig === 'object' && Object.keys(okapiConfig).length > 0)
-        ? { ...okapiConfig, ...parsedTenant } : { withoutOkapi: true };
+      const okapi = (typeof stripesOkapi === 'object' && Object.keys(stripesOkapi).length > 0)
+        ? { ...stripesOkapi, ...parsedTenant } : { withoutOkapi: true };
 
-      const initialState = merge({}, { okapi }, props.initialState);
+      const initialState = merge({}, { okapi, config: stripesConfig }, props.initialState);
 
-      this.logger = configureLogger(config);
+      this.config = stripesConfig;
+      this.branding = stripesBranding;
+      this.logger = configureLogger(stripesConfig);
       this.epics = configureEpics(connectErrorEpic);
       this.store = configureStore(initialState, this.logger, this.epics);
 
@@ -91,6 +147,8 @@ export default class StripesCore extends Component {
         modules: modulesInitialState,
       };
     } else {
+      this.config = localConfig;
+      this.branding = localBranding;
       this.state = {
         isStorageEnabled: false,
         actionNames: [],
@@ -102,12 +160,10 @@ export default class StripesCore extends Component {
   async componentDidMount() {
     if (this.state.isStorageEnabled) {
       try {
-        const modules = await getModules();
+        const modules = await getModules(this.config);
 
-        const folioConfig = await localforage.getItem(stripesHubAPI.FOLIO_CONFIG_KEY);
-        const brandingConfig = await localforage.getItem(stripesHubAPI.BRANDING_CONFIG_KEY);
         const discoveryUrl = await localforage.getItem(stripesHubAPI.DISCOVERY_URL_KEY);
-        const hostLocation = await localforage.getItem(stripesHubAPI.HOST_LOCATION_KEY);
+        const hostUrl = await localforage.getItem(stripesHubAPI.HOST_URL_KEY);
         const remotesList = await localforage.getItem(stripesHubAPI.REMOTE_LIST_KEY);
 
         const actionNames = gatherActions(modules);
@@ -116,10 +172,8 @@ export default class StripesCore extends Component {
           actionNames,
           modules,
           stripesHub: {
-            folioConfig,
-            brandingConfig,
             discoveryUrl,
-            hostLocation,
+            hostUrl,
             remotesList,
           }
         });
@@ -151,7 +205,7 @@ export default class StripesCore extends Component {
     const { initialState, ...props } = this.props;
 
     return (
-      <StrictWrapper>
+      <StrictWrapper config={this.config}>
         <div
           id={eventsPortal}
           className={css.eventsContainer}
@@ -160,10 +214,11 @@ export default class StripesCore extends Component {
           store={this.store}
           epics={this.epics}
           logger={this.logger}
-          config={config}
+          config={this.config}
+          branding={this.branding}
           actionNames={actionNames}
           modules={modules}
-          disableAuth={(config?.disableAuth) || false}
+          disableAuth={(this.config?.disableAuth) || false}
           stripesHub={stripesHub}
           {...props}
         />