+++ src/components/Root/Root.js
@@ -9,7 +9,7 @@ import { QueryClientProvider } from 'react-query';
 import { ApolloProvider } from '@apollo/client';
 
 import { ErrorBoundary } from '@folio/stripes-components';
-import { branding, metadata, icons as configIcons } from 'stripes-config';
+import { metadata, icons as configIcons } from 'stripes-config';
 
 import { ConnectContext } from '@folio/stripes-connect';
 import initialReducers from '../../initialReducers';
@@ -31,7 +31,6 @@ import {
   rotationHandler,
   ResetTimer
 } from './token-util';
-import { getStripesHubConfig } from './stripes-hub-util';
 import { modulesInitialState } from '../../ModulesContext';
 import {
   RTR_FLS_TIMEOUT_EVENT,
@@ -116,6 +115,7 @@ class Root extends Component {
    */
   async componentDidMount() {
     const { okapi, store, defaultTranslations } = this.props;
+
     if (this.withOkapi) {
       // check for an existing session in storage. if found, initialize the
       // end-of-session timers. for new sessions, this happens in LoginCtrl
@@ -125,9 +125,14 @@ class Root extends Component {
         await this.handleRotation(sess.tokenExpiration);
       }
     }
-    const locale = this.props.config.locale ?? 'en-US';
-    // TODO: remove this after we load locale and translations at start from a public endpoint
-    loadTranslations(store, locale, defaultTranslations);
+
+    // checkOkapiSession triggers loadTranslations for authenticated sessions
+    const { locale: currentLocale, translations } = store.getState().okapi;
+    if (!translations) {
+      const locale = currentLocale || this.props.config.locale || 'en-US';
+      // TODO: remove this after we load locale and translations at start from a public endpoint
+      await loadTranslations(store, locale, defaultTranslations);
+    }
     this.updateQueryResourceStateKey();
   }
 
@@ -183,6 +188,7 @@ class Root extends Component {
       store,
       epics,
       config,
+      branding,
       okapi,
       actionNames,
       token,
@@ -224,18 +230,20 @@ class Root extends Component {
     //   render runs when props change. realistically, that'll never happen
     //   since config values are read only once from a static file at build
     //   time, but still, props are props so technically it's possible.
-    config.rtr = configureRtr(stripesHub?.folioConfig?.rtr || this.props.config.rtr);
+    config.rtr = configureRtr(this.props.config.rtr);
 
     // if we have a stripesHub configuration, pass it to stripes...
-    const { stripesOkapi, stripesConfig, stripesBranding } = getStripesHubConfig(okapi, config, branding, stripesHub);
+    okapi.discoveryUrl = stripesHub.discoveryUrl;
+    okapi.hostUrl = stripesHub.hostUrl;
+    okapi.remotesList = stripesHub.remotesList;
 
     const stripes = new Stripes({
       logger,
       store,
       epics,
-      config: stripesConfig,
-      okapi: stripesOkapi,
-      branding: stripesBranding,
+      config,
+      okapi,
+      branding,
       withOkapi: this.withOkapi,
       setToken: (val) => { store.dispatch(setOkapiToken(val)); },
       setIsAuthenticated: (val) => { store.dispatch(setIsAuthenticated(val)); },
@@ -325,6 +333,7 @@ Root.propTypes = {
   plugins: PropTypes.object,
   bindings: PropTypes.object,
   config: PropTypes.object,
+  branding: PropTypes.object,
   okapi: PropTypes.shape({
     url: PropTypes.string,
     tenant: PropTypes.string,